[Feature] Create combined user test file
